+++ SBO-Lib/lib/SBO/App.pm
@@ -13,7 +13,7 @@ use strict;
 use warnings FATAL => 'all';
 use File::Basename;
 
-our $VERSION = '4.1.2';
+our $VERSION = '4.1.3';
 
 sub new {
   my $class = shift;